requirements.txt
tensorflow==2.15.0
fastapi==0.104.1
uvicorn==0.24.0
pillow==10.1.0
numpy==1.24.3
matplotlib==3.8.2
scikit-learn==1.3.2
pandas==2.1.3
opencv-python==4.8.1.78
tqdm==4.66.1